requirements.txt
geopy==2.1.0
Jinja2==2.11.2
PyYAML==5.3.1
requests==2.25.1
urllib3==1.26.2
